+++ tests/pg/AuthenticateRealarmRest.php
@@ -4,15 +4,13 @@
  * Access Token 요청 예제입니다.
  */
 require_once '../vendor/autoload.php';
+require_once __DIR__ . '/../config.php';
 
 // require_once __DIR__.'/../src/BootpayApi.php';
 
 use Bootpay\ServerPhp\BootpayApi;
 
-BootpayApi::setConfiguration(
-    '5b8f6a4d396fa665fdc2b5ea',
-    'rm6EYECr6aroQVG2ntW0A6LpWnkTgP4uQ3H18sDDUYw='
-);
+setupActiveBootpayApi();
 
 $token = BootpayApi::getAccessToken();
 var_dump($token);